improve definition adding, and package structures
