+++ src/main/java/com/emarte/regurgitator/core/AllMatchesBehaviour.java
@@ -1,18 +1,15 @@
-package com.emarte.regurgitator.core.behaviour.rules;
-
-import com.emarte.regurgitator.core.*;
+package com.emarte.regurgitator.core;
 
 import java.util.*;
 
-public class FirstMatchBehaviour implements RulesBehaviour {
-	private Log log = Log.getLog(FirstMatchBehaviour.class);
+public class AllMatchesBehaviour implements RulesBehaviour {
+	private Log log = Log.getLog(AllMatchesBehaviour.class);
 
 	@Override
 	public List<Object> evaluate(Object decisionId, List<Object> evaluatedStepIds, List<Object> allStepIds, Object defaultStepId) {
 		if(evaluatedStepIds.size() > 0) {
-			Object firstMatch = evaluatedStepIds.get(0);
-			log.debug("Returning first rule match '" + firstMatch + "'");
-			return Arrays.asList(firstMatch);
+			log.debug("Returning all rule matches '" + evaluatedStepIds + "'");
+			return evaluatedStepIds;
 		}
 
 		if(defaultStepId != null) {
