+++ src/main/java/com/emarte/regurgitator/core/EqualsBehaviour.java
@@ -1,6 +1,4 @@
-package com.emarte.regurgitator.core.behaviour.condition;
-
-import com.emarte.regurgitator.core.*;
+package com.emarte.regurgitator.core;
 
 public class EqualsBehaviour implements ConditionBehaviour {
 	private static Log log = Log.getLog(EqualsBehaviour.class);
@@ -11,7 +9,8 @@ public class EqualsBehaviour implements ConditionBehaviour {
 		Parameter parameter = message.getContextValue(location);
 
 		if(parameter != null) {
-			equals = parameter.getValue().equals(parameter.getType().convert(conditionValue));
+			ParameterType parameterType = parameter.getType();
+			equals = parameter.getValue().equals(parameterType.convert(conditionValue));
 		}
 
 		log.debug("Parameter '" + location + (equals ? "' equals" : "' does not equal") + " value '" + conditionValue + "'");
