+++ src/main/java/com/emarte/regurgitator/core/RulesBehaviour.java
@@ -1,6 +1,6 @@
 package com.emarte.regurgitator.core;
 
-import java.util.*;
+import java.util.List;
 
 public interface RulesBehaviour {
 	public List<Object> evaluate(Object decisionId, List<Object> evaluatedStepIds, List<Object> allStepIds, Object defaultStepId);
