+++ src/main/java/com/emarte/regurgitator/core/UuidGenerator.java
@@ -1,6 +1,4 @@
-package com.emarte.regurgitator.core.generator;
-
-import com.emarte.regurgitator.core.*;
+package com.emarte.regurgitator.core;
 
 import java.util.UUID;
