+++ src/test/java/com/emarte/regurgitator/core/ConditionBehaviourTest.java
@@ -1,12 +1,11 @@
 package com.emarte.regurgitator.core;
 
-import com.emarte.regurgitator.core.behaviour.condition.*;
 import org.junit.Test;
 
 import java.util.Arrays;
 
 import static com.emarte.regurgitator.core.ConflictPolicy.REPLACE;
-import static com.emarte.regurgitator.core.type.CoreTypes.*;
+import static com.emarte.regurgitator.core.CoreTypes.*;
 import static junit.framework.Assert.*;
 
 public class ConditionBehaviourTest {
