+++ src/test/java/com/emarte/regurgitator/core/IdentifySessionTest.java
@@ -3,7 +3,7 @@ package com.emarte.regurgitator.core;
 import org.junit.Test;
 
 import static com.emarte.regurgitator.core.ConflictPolicy.REPLACE;
-import static com.emarte.regurgitator.core.type.CoreTypes.STRING;
+import static com.emarte.regurgitator.core.CoreTypes.STRING;
 import static junit.framework.Assert.assertEquals;
 
 public class IdentifySessionTest {
